requirements.txt
pip~=23.3.1
wheel~=0.38.4
setuptools~=65.5.1
python-dotenv~=1.0.0
flask~=3.0.0
torch~=2.1.1
transformers~=4.35.2
psutil~=5.9.6
accelerate~=0.25.0
scipy~=1.11.4
